Update Gradle and Spring Boot versions
The Gradle distribution version has been upgraded from 7.6 and 7.5.1 to 8.5 for the reactive-customers, reactive-officers, and restclient modules. Furthermore, Spring Boot's version is updated from 3.0.5, 3.0.2, and 2.7.3 to 3.2.1 in their respective build.gradle files. This ensures the applications are using the latest tools and dependencies, which can provide updated features and improved performance.