Update Spring Boot version and refactor tests
This commit updates the version of Spring Boot used in the build.gradle file. Additionally, we've simplified the syntax of server response expectations in CustomerControllerTest to make the assertions more readable. We also removed the unnecessary print statement in the deleteCustomer test in CustomerRepositoryTest file. Lastly, a log() call was added in the findById method in the CustomerController for debugging purposes.